Commit 6 (passed evrything, need to clean up everything and explain code next.)

+++ src/main/java/DAO/MessageDAO.java
@@ -121,72 +121,26 @@ public class MessageDAO {
       }
       return null;
   }
-
-  public Message deleteMessageById(Message message) throws Exception
-{
+  public Message deleteMessageById(int message_id) throws Exception {
     Connection connection = ConnectionUtil.getConnection();
-    String sql = "DELETE FROM message WHERE message_id=?";
-    try
-    {
-        PreparedStatement preparedStatement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-        int id = message.getMessage_id();
-        preparedStatement.setInt(1, id);
-        preparedStatement.executeUpdate();
-        return new Message(message.getMessage_id(), message.getPosted_by(), message.getMessage_text(), message.getTime_posted_epoch());
+    Message message = getMessageById(message_id);
+
+    if (message == null) {
+        return null;
     }
-    catch (RuntimeException | SQLException runtimeException) 
-    {
+
+    String sql = "DELETE FROM message WHERE message_id=?";
+    try {
+        PreparedStatement preparedStatement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
+        preparedStatement.setInt(1, message_id);
+        preparedStatement.executeUpdate();
+        return message;
+    } catch (RuntimeException | SQLException runtimeException) {
         System.err.println("MessageDAO::deleteObject, RuntimeException occurred, message follows.");
         System.err.println(runtimeException);
         throw runtimeException;
     }
 }
-
-
-//   public Message deleteMessageById(Message message){
-//     Connection connection = ConnectionUtil.getConnection();
-//     try {
-//         //Write SQL logic here. When inserting, you only need to define the 
-//         String sql = "DELETE FROM message WHERE message_id = ?;" ;
-//         PreparedStatement preparedStatement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-
-//         //write preparedStatement's setString and setInt methods here.
-//         int id = message.getMessage_id();
-//         preparedStatement.setInt(1, id);
-
-//         preparedStatement.executeUpdate();
-//         ResultSet pkeyResultSet = preparedStatement.getGeneratedKeys();
-//         if(pkeyResultSet.next()){
-//             int message_id_to_delete = (int) pkeyResultSet.getInt(1);
-//             message.setMessage_id(message_id_to_delete);
-//           //System.out.println(new Message(message.getMessage_id(), message.getPosted_by(), message.getMessage_text(), message.getTime_posted_epoch()));
-//           return new Message(message.getMessage_id(), message.getPosted_by(), message.getMessage_text(), message.getTime_posted_epoch());
-//         }
-//     }catch(SQLException e){
-//         System.out.println(e.getMessage());
-//     }
-//     return null;
-// }
-
-// public Message deleteMessageById(int id){
-//     Connection connection = ConnectionUtil.getConnection();
-//     try {
-//         String sql = "DELETE FROM message WHERE message_id = ?;";
-        
-//         PreparedStatement preparedStatement = connection.prepareStatement(sql);
-//         preparedStatement.setInt(1, id);
-
-//         ResultSet rs = preparedStatement.executeQuery();
-//         while(rs.next()){
-//             Message message_by_id = new Message(rs.getInt("message_id"), rs.getInt("posted_by"),
-//                                     rs.getString("message_text"), rs.getLong("time_posted_epoch"));
-//             return message_by_id;
-//         }
-//     } catch(SQLException e){
-//         System.out.println(e.getMessage());
-//     }
-//     return null;
-// }
 
 
 public void updateMessage(int id, String message) {
@@ -202,79 +156,6 @@ public void updateMessage(int id, String message) {
         System.out.println(e.getMessage());
     }
 }
-
-
-
-
-//    public List<Message> updateMessage(int id, String message){
-//     Connection connection = ConnectionUtil.getConnection();
-//     List<Message> messages = new ArrayList<>();
-//     try {
-//         //Write SQL logic here
-//         String sql = "UPDATE message SET message_text =? WHERE message_id = ?;";
-//         PreparedStatement preparedStatement = connection.prepareStatement(sql);
-
-//         //int posted_by = message.getPosted_by();
-//         //String message_text = message.getMessage_text();
-//         //long time_posted = message.getTime_posted_epoch();
-
-
-//         preparedStatement.setInt(2,id);
-//         //preparedStatement.setInt(1,posted_by);
-//         preparedStatement.setString(1,message);
-//         //preparedStatement.setLong(3,time_posted);
-//         //System.out.println(preparedStatement);
-
-//         //PreparedStatement preparedStatement = connection.prepareStatement(sql);
-//         ResultSet rs = preparedStatement.executeQuery();
-//         while(rs.next()){
-//             Message message1 = new Message(rs.getInt("message_id"), rs.getInt("posted_by"),
-//                     rs.getString("message_text"), rs.getLong("time_posted_epoch"));
-//             messages.add(message1);
-//             //System.out.println(message1);
-//         }
-//     }catch(SQLException e){
-//         System.out.println(e.getMessage());
-//     }
-//     return messages;
-// }
-
-//    public void updateMessage(int id, Message message){
-//       Connection connection = ConnectionUtil.getConnection();
-//       try {
-//           //Write SQL logic here
-          
-//           String sql = "UPDATE message SET posted_by =?, message_text =?, time_posted_epoch =? WHERE message_id = ?;";
-//           PreparedStatement preparedStatement = connection.prepareStatement(sql);
-
-//           //write PreparedStatement setString and setInt methods here.
-
-        //   int posted_by = message.getPosted_by();
-        //   String message_text = message.getMessage_text();
-        //   long time_posted = message.getTime_posted_epoch();
-
-
-        //   preparedStatement.setInt(4,id);
-        //   preparedStatement.setInt(1,posted_by);
-        //   preparedStatement.setString(2,message_text);
-        //   preparedStatement.setLong(3,time_posted);
-
-//         //   ResultSet rs = preparedStatement.executeQuery();
-//         //     while(rs.next()){
-//         //         String generated_message = (String) rs.getString(1);
-//         //         message.setMessage_text(generated_message);
-//         //         message = new Message(rs.getInt("message_id"), rs.getInt("posted_by"), rs.getString("message_text"),
-//         //                 rs.getLong("time_posted_epoch"));
-//         //         return message;
-//         //     }
-
-
-//           preparedStatement.executeUpdate();
-//       }catch(SQLException e){
-//           System.out.println(e.getMessage());
-//       }
-//   }
-
